feat: move to view contract

+++ contracts/src/view/TankGameView.sol
@@ -14,25 +14,18 @@ contract TankGameView {
     struct TankLocation {
         TankGame.Tank tank;
         ITankGame.Point position;
-        uint tankId;
+        uint256 tankId;
     }
 
     function getAllTanks() external view returns (TankLocation[] memory) {
         TankLocation[] memory tanksWithLocation = new TankLocation[](
             tankGame.playersCount()
         );
-        for (uint i = 1; i <= tankGame.playersCount(); i++) {
-            (uint x, uint y) = tankGame.tankToPosition(i);
+        for (uint256 i = 1; i <= tankGame.playersCount(); i++) {
+            (uint256 x, uint256 y) = tankGame.tankToPosition(i);
             ITankGame.Point memory position = ITankGame.Point(x, y);
-            (address owner, uint hearts, uint aps, uint range) = tankGame.tanks(
-                i
-            );
-            TankGame.Tank memory tank = TankGame.Tank(
-                owner,
-                hearts,
-                aps,
-                range
-            );
+            (address owner, uint256 hearts, uint256 aps, uint256 range) = tankGame.tanks(i);
+            TankGame.Tank memory tank = TankGame.Tank(owner, hearts, aps, range);
             tanksWithLocation[i - 1] = TankLocation(tank, position, i);
         }
         return tanksWithLocation;
